changed amount data types to decimal

+++ db/migrate/07_add_remaining_amount.rb
@@ -1,5 +1,5 @@
 class AddRemainingAmount < ActiveRecord::Migration[5.2]
   def change
-    add_column :budgets, :remaining_amount, :integer
+    add_column :budgets, :remaining_amount, :numeric
   end
 end
